refactor: 도메인 로직과 관련된 객체를 util에서 domain 패키지로 변경
